docs: :fire: remove "correctness" from overall aim
"correctness" can mean "being in agreement with the true facts" (according to Cambridge dict) which is outside of the scope of this package; we don't ensure that what is in the descriptor is actually correct or true. According to Merriam-Webster, "correctbess" can mean "conforming to an approved or conventional standard" and with this definition its basically the same as "compliance". So, it's either not entirely what we want with this package or very similar to compliance, so I'll remove it here.