docs: :memo: add combined C4 component and code diagram (#154)
# Description

This PR adds a combined C4 component/code diagram for the core Python
package. It kind of mirrors the existing flow diagram in the interface
docs. I initially made a more detailed version trying to mirror the
implementation more closely, but for the purpose of this, I think this
might be fine with this level of detail.

Related to #46 

Needs an in-depth review.

## Checklist

- [X] Formatted Markdown
- [X] Ran `just run-all`