feat: :sparkles: add `RequiredCheck` (#122)
# Description

This PR handles the special case of requiring a field to be non-null by
subclassing `CustomCheck` to create `RequiredCheck`.

I thought this was a nice way of splitting out the logic for required
rules and that required rules were special enough to warrant their own
class.

I also thought it made sense to limit what kind of JSON paths the
required rule can be applied to. It makes sense to mark a dict property
(e.g. `$.resources[*].title`) as required because it is very obvious
what it means for this field to be missing.
- But marking an array item itself (e.g. `$.resources[*]` or
`$.resources[2]`) as required makes less sense to me. "I want the second
resource to exist in particular" is a weird constraint. If someone cares
about the number of resources, they should be checking the length of the
array.
- Same for "vague" JSON paths like `$..title`: selecting all `title`
fields under the root node and then checking if they exist feels
circular and unnecessary.

:arrow_right: So I restricted `RequiredCheck` to sensible paths only,
which also made the code simpler.

Closes #120 

Needs an in-depth review.

## Checklist

- [x] Formatted Markdown
- [x] Ran `just run-all`

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>