refactor: :recycle: handle grouped errors without modifying input (#177)
# Description
This is a refactor of how the functions that handle grouped errors
handle those grouped errors.

Previously they took in all `schema_errors` as input, added and removed
errors, and returned a modified list.

The more functions we have, the messier this becomes, so this PR
refactors the functions to return a list of errors to add and a list of
errors to remove. Then, we can just do all additions and removals once
at the end.

Needs an in-depth review.

## Checklist

- [x] Formatted Markdown
- [x] Ran `just run-all`