feat: :sparkles: handle grouped errors under resource fields (#175)
# Description

This PR adds a function for handling grouped errors under
`$.resources[x].schema.fields[x]`.
Here, the problem comes from the fact that each field type has its own
sub-JSON-schema, and each one of these sub-schemas flags issues when the
type of a field is not its own type. So, if a field has `type="number"`
and there is something wrong with the field, then the sub-schemas for
`year`, `string`, etc. will also flag issues. The goal is to flag issues
only for `number`.

Part of #15 

Needs an in-depth review.

## Checklist

- [x] Formatted Markdown
- [x] Ran `just run-all`