feat: :sparkles: handle grouped foreign key errors (#188)
# Description

This PR handles grouped errors under
`$.resources[x].schema.foreignKeys`.

The situation here is that a foreign key can be specified either as a
string or an array of strings. The target that the foreign key points to
(usually a primary key) must then also have the same type. So we can
look at what type the foreign key is and keep only errors flagged by
that sub-schema.

There are some similarities with the code for primary keys -- maybe we
can pull some helpers out later.

Note that there are a number of [`MUST` constraints on foreign
keys](https://datapackage.org/standard/table-schema/#foreignKeys) that
are not actually included in the JSON schema. Tracking them in #187

Part of #15 

Needs an in-depth review.

## Checklist

- [x] Formatted Markdown
- [x] Ran `just run-all`

Co-authored-by: Luke W. Johnston <lwjohnst86@users.noreply.github.com>