feat: :sparkles: support union operator in `RequiredCheck` (#201)
# Description

This PR adds support for the union JSON path operator for
`RequiredCheck`s.

To be able to do this more intelligently than a monster regex, I looked
into how `jsonpath` represents a JSON path internally and used their
terms/classification. This also allowed me to be more precise about what
JSON paths we are not allowing for `RequiredCheck`s.

Closes #193 

Needs an in-depth review.

## Checklist

- [x] Formatted Markdown
- [x] Ran `just run-all`

---------

Co-authored-by: Luke W. Johnston <lwjohnst86@users.noreply.github.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>