fix: Only add the error instance if it is hashable
Since we perform deduplication of Issues later,
we need to either only allow hashable items
or write custome comparison logic
to compare unhashable ones.
Technically,
writing custom comparison logic
that ignores unhashable values
might be the most theoretically appropriate
(since it is redundant to compare `instance`
when also comparing `jsonpath`),
but it could look convoluted
and have worse performance.
This is a simpler solution
that is easy to understand
and that does not involve any significant performance penalty
as the extra comparison is of small items.
Trying this out
as a first approach
and changing if we notice that there are occasions
where we would value having access to an unhashable instance.
For more discussion, see:
https://github.com/seedcase-project/check-datapackage/pull/208#discussion_r2555916842