feat: :sparkles: display helpful error message for empty input (#243)
# Description

Currently the output when an empty list is passed to `check` is not that
helpful. I'm sure the error message can be improved further, but this is
just to get the ball rolling since there is synergy with #241

```sh
uv run python -c 'import check_datapackage as cdp; cdp.check([], error=True)'              
```

Output before this PR (after merging #241):

```  
1 issue was found in your `datapackage.json`:

At :
|
|   []
|   ^^
[] is not of type 'object'
```

Output after this PR:

```
`check()` requires a dictionary with metadata, but received an empty list.
```
Closes https://github.com/seedcase-project/check-datapackage/issues/220

Needs a quick review.

## Checklist

- [ ] Formatted Markdown
- [x] Ran `just run-all`

---------

Co-authored-by: martonvago <57952344+martonvago@users.noreply.github.com>
Co-authored-by: Luke W. Johnston <lwjohnst86@users.noreply.github.com>