docs: :lipstick: improve code output format and color in docs (#260)
# Description

Needs a quick review.

As mentioned in #259, there are a few issues with the default rich
representation of the output: the defaults colors don't look that great
on a white background, the lack of soft-wrap is sometimes and issue, and
there is also an extra newline inserted after each code block before the
output appears.

## Checklist

- [x] Ran `just run-all`

---------

Co-authored-by: Luke W. Johnston <lwjohnst86@users.noreply.github.com>