feat: :sparkles: add env variable to help debug issues (#223)
# Description

I think we need to go through many types of error messages manually and
see if the way the error is presented via `explain` is helpful. This
debug option helps with that by always printing the properties dict, its
issues, and the output from `explain`. It also enables us to easily do
this on all test messages.

This is now easy to detect by running the test suite with the debug
option enabled, but we still need to make decision about what to do with
these errors by opening a new issue for each one of them and discuss the
best solutions.

Needs an in-depth review.

---------

Co-authored-by: Luke W. Johnston <lwjohnst86@users.noreply.github.com>